feat: added fsd arch
