+++ src/app/routes/index.ts
@@ -1,0 +1,9 @@
+import { createRouter, createWebHistory } from 'vue-router'
+import { routes } from './routes'
+
+const router = createRouter({
+  history: createWebHistory(import.meta.env.BASE_URL),
+  routes
+})
+
+export default router
