+++ src/pages/about/index.vue
@@ -1,0 +1,1 @@
+<template>About</template>
