+++ src/pages/home/index.vue
@@ -1,0 +1,1 @@
+<template>Home</template>
